+++ app/models/check_group.rb
@@ -3,8 +3,4 @@ class CheckGroup < ApplicationRecord
   has_many :disclosure_checks, dependent: :destroy
 
   scope :with_completed_checks, -> { joins(:disclosure_checks).where(disclosure_checks: { status: :completed }).distinct }
-
-  def multiple_sentences?
-    disclosure_checks.size > 1
-  end
 end
